Split Album page template and load function
